Added error suppression support for XML,re:View,Latex and Wiki comment formats
